Add flag to originalEvent to show event is simulated.
Touch Punch changes the order of touch and mouse events fired, firing a "mouseenter" event before the "touchstart" event.  This makes determining mouse/touch priority difficult when defining behaviours.

Adding a "simulated" flag to the originalEvent means that trivial detection can be done to determine if the event raised is a natural mouse event or one raised by Touch Punch.